\name{kolmogorov}
\alias{kolmogorov}
\title{
Evaluates the KS statistic for some given data.
}
\description{
This function evaluates the KS statistic of some given data for the specified random variable
using an approximation of the CDF on the specified interval.
}
\usage{
 kolmogorov(X, min, max, N, samples)
}
\arguments{
 \item{X}{the random variable.}
 \item{min}{the lower bound.}
 \item{max}{the upper bound.}
 \item{N}{the number of bins.}
 \item{samples}{the samples.}
}
